+++ include/ignition/gazebo/components/JointPositionReset.hh
@@ -33,10 +33,12 @@ inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
   /// \brief Joint positions in SI units (rad for revolute, m for prismatic).
+  ///
   /// The component wraps a std::vector of size equal to the degrees of freedom
   /// of the joint.
-  using JointPositionReset = Component<std::vector<double>, class JointPositionResetTag,
-                                  serializers::VectorDoubleSerializer>;
+  using JointPositionReset = Component<std::vector<double>,
+                                       class JointPositionResetTag,
+                                       serializers::VectorDoubleSerializer>;
   IGN_GAZEBO_REGISTER_COMPONENT(
       "ign_gazebo_components.JointPositionReset", JointPositionReset)
 }