+++ src/Server_TEST.cc
@@ -64,6 +64,7 @@ TEST_P(ServerFixture, DefaultServerConfig)
   EXPECT_FALSE(serverConfig.LogRecordPath().empty());
   EXPECT_FALSE(serverConfig.LogIgnoreSdfPath());
   EXPECT_TRUE(serverConfig.LogPlaybackPath().empty());
+  EXPECT_TRUE(serverConfig.LogRecordCompressPath().empty());
   EXPECT_EQ(0u, serverConfig.Seed());
   EXPECT_EQ(123ms, serverConfig.UpdatePeriod().value_or(123ms));
   EXPECT_TRUE(serverConfig.ResourceCache().empty());
@@ -273,11 +274,14 @@ TEST_P(ServerFixture, ServerConfigLogRecord)
   auto logPath = common::joinPaths(
       std::string(PROJECT_BINARY_PATH), "test_log_path");
   auto logFile = common::joinPaths(logPath, "state.tlog");
+  auto compressedFile = logPath + ".zip";
 
   igndbg << "Log path [" << logPath << "]" << std::endl;
 
   common::removeAll(logPath);
+  common::removeAll(compressedFile);
   EXPECT_FALSE(common::exists(logFile));
+  EXPECT_FALSE(common::exists(compressedFile));
 
   {
     gazebo::ServerConfig serverConfig;
@@ -291,6 +295,38 @@ TEST_P(ServerFixture, ServerConfigLogRecord)
   }
 
   EXPECT_TRUE(common::exists(logFile));
+  EXPECT_FALSE(common::exists(compressedFile));
+}
+
+/////////////////////////////////////////////////
+TEST_P(ServerFixture, ServerConfigLogRecordCompress)
+{
+  auto logPath = common::joinPaths(
+      std::string(PROJECT_BINARY_PATH), "test_log_path");
+  auto logFile = common::joinPaths(logPath, "state.tlog");
+  auto compressedFile = logPath + ".zip";
+
+  igndbg << "Log path [" << logPath << "]" << std::endl;
+
+  common::removeAll(logPath);
+  common::removeAll(compressedFile);
+  EXPECT_FALSE(common::exists(logFile));
+  EXPECT_FALSE(common::exists(compressedFile));
+
+  {
+    gazebo::ServerConfig serverConfig;
+    serverConfig.SetUseLogRecord(true);
+    serverConfig.SetLogRecordPath(logPath);
+    serverConfig.SetLogRecordCompressPath(compressedFile);
+
+    gazebo::Server server(serverConfig);
+    EXPECT_EQ(0u, *server.IterationCount());
+    EXPECT_EQ(3u, *server.EntityCount());
+    EXPECT_EQ(4u, *server.SystemCount());
+  }
+
+  EXPECT_FALSE(common::exists(logFile));
+  EXPECT_TRUE(common::exists(compressedFile));
 }
 
 /////////////////////////////////////////////////